feat: 상품 삭제 시 ES 도큐먼트 동기화 및 ProductCdcEvent 제거
- ProductDeletedEvent를 통한 ES 도큐먼트 삭제 흐름 추가
- 미사용 ProductCdcEvent를 ProductUpdatedEvent로 대체
